refactor: :art: Alpha Channel now supports floating point

+++ src/Anweshan/Image/Manipulators/Helpers/Color.php
@@ -420,7 +420,7 @@ class Color {
 			
 			if (preg_match(self::SHORT_ARGB, $value)) {
 				$rgba = $this->parseHex(substr($value, 1).substr($value, 1));
-				$alpha = substr($value, 0, 1) / 10;
+				$alpha = (float) substr($value, 0, 1) / 10;
 				break;
 			}
 			
@@ -433,7 +433,7 @@ class Color {
 			
 			if (preg_match(self::LONG_ARGB, $value)) {
 				$rgba = $this->parseHex(substr($value, 2));
-				$alpha = substr($value, 0, 2) / 100;
+				$alpha = (float) substr($value, 0, 2) / 100;
 				$default = false;
 				break;
 			}
